TextEditor, TextLineRenderer: partial support for word wrapping
Line number position is now based on the baseline of the first line of a text renderer, even for single lines

Supported in read-only mode only. Will be buggy if editor is editable. For instance, selection with keyboard does not always work correctly because it cannot account for text line renderers that are taller than a single line.